# back-end/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Flipic API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:4000",
  "basePath": "/",
  "tags": [
    {
      "name": "Quiz",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "any description..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/quiz": {
      "post": {
        "tags": [
          "Quiz"
        ],
        "description": "Endpoint to create a quiz",
        "parameters": [
          {
            "name": "quiz",
            "in": "body",
            "description": "Quiz information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Quiz"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "put": {
        "tags": [
          "Quiz"
        ],
        "description": "Endpoint to edit a quiz",
        "parameters": [
          {
            "name": "quiz",
            "in": "body",
            "description": "Quiz information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Quiz"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/quiz/{id}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "description": "Endpoint to get a quiz",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "delete": {
        "tags": [
          "Quiz"
        ],
        "description": "Endpoint to delete a quiz",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/quiz-results/{id}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "description": "Endpoint to get a quizs results",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/quiz-matchup/{id}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "description": "Endpoint to get a matchup of 2 items",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/quiz-vote": {
      "post": {
        "tags": [
          "Quiz"
        ],
        "description": "Endpoint to vote on a matchup",
        "parameters": [
          {
            "name": "quiz",
            "in": "body",
            "description": "Quiz and item Ids.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/QuizItemIds"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/quiz-all": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "description": "Endpoint to get all quiz",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    }
  },
  "definitions": {
    "Item": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Item name"
        },
        "icon": {
          "type": "string",
          "example": "Icon name"
        }
      }
    },
    "Quiz": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Quiz Name"
        },
        "quizCreator": {
          "type": "string",
          "example": "Quiz Creators Name"
        },
        "privacyType": {
          "type": "string",
          "example": "Public"
        },
        "createdBy": {
          "type": "string",
          "example": "Created By"
        },
        "audienceSize": {
          "type": "string",
          "example": "5"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Item"
          }
        }
      }
    },
    "QuizItemIds": {
      "type": "object",
      "properties": {
        "quizId": {
          "type": "string",
          "example": "Quiz Id"
        },
        "itemId": {
          "type": "string",
          "example": "Item Id"
        }
      }
    }
  }
}